Merge RenderZenith 1.3.0 #5
* trace/trace_sched_wakeup: Fix compliation error when CONFIG_SCHED_TRACER=y

Signed-off-by: RenderBroken <zkennedy87@gmail.com>

* sched/fair.c: Don't allow SchedTune boosted tasks to be migrated to small cores

We want boosted tasks to run on big cores. But CAF's load balancer
changes do not account for SchedTune boosting, so this allows for
boosted tasks to be migrated to a suboptimal core. Let's mitigate
this by setting the LBF_IGNORE_BIG_TASKS for tasks migrating from a
larger capacity core to a smaller one and to check if the task is
SchedTune boosted. If both are true, do not migrate this task.

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>

* Revert "sched: Enable TTWU_QUEUE as a sched feature for SD845."

This reverts commit b6ef2621a19dfd8b73e437718c53a11b6ca5187f.

* sched/fair: Fix redundant load balancer reattempt due to LBF_ALL_PINNED

LBF_ALL_PINNED flag should cleared in can_migrate_task() if the task
can run on the destination CPU during load balance. In current code,
can_migrate_task() return incorrectly without clearing this flag
in case if the task can't be migrated to the destination CPU due to
cumulative window demand constraints. Since LBF_ALL_PINNED flag
is not cleared, load balancer thinks that none of the tasks running
on the busiest group can't be migrated to the destination CPU due
to affinity settings and tries to find another busiest group. Prevent
this incorrect reattempt of load balance by clearing LBF_ALL_PINNED
flag right after the task affinity check in can_migrate_task().

Change-Id: Iad1cf42b1aaf70106ee5ecfbd9499ccb6eb7497e
[clingutla@codeaurora.org: Resolved merge conflicts]
Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>